Refactor: use cb::net:: for sockets
This removes duplication of functionality from cbsocket and
we can reduce the number of #ifdefs to handle the difference
between *nix and Windows

Reviewed-on: http://review.couchbase.org/90784
Tested-by: Build Bot <build@couchbase.com>

+++ daemon/connection_mcbp.h
@@ -284,7 +284,7 @@ public:
      * @param m the message header to send
      * @return the number of bytes sent, or -1 for an error
      */
-    int sendmsg(struct msghdr* m);
+    ssize_t sendmsg(struct msghdr* m);
 
     enum class TransmitResult {
         /** All done writing. */
